refactor: logic for speaker layout only
